+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Unreleased
 
 * Add focus highlight indicator for `Scroll!` with no focusable children.
+* Change `Button::cmd` defaults.
+    - No longer sets `on_disabled_click`.
 * Change `toggle::RadioStyle!`.
     - Align mark to top in multi-line content.
 * Change `toggle::CheckStyle!`.